+++ test/novident_nodes_test.dart
@@ -24,6 +24,53 @@ void main() {
     expect(node.first.level, 1);
   });
 
+  test('should move first child node into the last child one', () {
+    final DirectoryNode node = DirectoryNode(
+      details: NodeDetails.zero(),
+      children: <Node>[
+        // this node
+        FileNode(
+          details: NodeDetails.byId(id: 'test', level: 0),
+          content: '',
+          name: 'File 1',
+        ),
+        DirectoryNode(
+          details: NodeDetails.byId(id: 'test 2', level: 0),
+          children: <Node>[
+            FileNode(
+              details: NodeDetails.byId(id: 'test 3', level: 0),
+              content: '',
+              name: 'File 2',
+            ),
+            // should be moved here
+          ],
+          name: 'Dir 2',
+        ),
+      ],
+      name: 'Dir',
+    );
+    expect(node.last.id, 'test 2');
+    expect(node.last.castToDir.first.castToFile.name, 'File 2');
+    expect(node.last.castToDir.length, 1);
+    final NodeContainer parent = node.elementAt(1) as NodeContainer;
+    if (Node.canMoveTo(node: node.first, target: parent)) {
+      final bool moved = Node.moveTo(
+        node: node.first,
+        newOwner: parent,
+      );
+      expect(moved, isTrue);
+    }
+    // root checks
+    expect(node.first.id, 'test 2');
+    expect(node.castToDir.length, 1);
+    // children checks
+    expect(node.first.castToDir.length, 2);
+    expect(node.first.castToDir.first.id, 'test 3');
+    expect(node.first.castToDir.last.id, 'test');
+    expect(node.first.castToDir.last.castToFile.name, 'File 1');
+    expect(node.first.castToDir.last.level, 2);
+  });
+
   test('should update node into correctly', () {
     final DirectoryNode node = DirectoryNode(
       details: NodeDetails.zero(),