+++ test/test_nodes/file_node.dart
@@ -57,42 +57,12 @@ class FileNode extends Node {
   }
 
   @override
-  int countAllNodes({required Predicate countNode}) {
-    return countNode(this) ? 1 : 0;
-  }
-
-  @override
-  int countNodes({required Predicate countNode}) {
-    return countNode(this) ? 1 : 0;
-  }
-
-  @override
-  bool deepExist(String id) {
-    return this.id == id;
-  }
-
-  @override
-  bool exist(String id) {
-    return this.id == id;
-  }
-
-  @override
   Map<String, dynamic> toJson() {
     return <String, dynamic>{
       'name': name,
       'details': details.toJson(),
       'content': content,
     };
-  }
-
-  @override
-  Node? visitAllNodes({required Predicate shouldGetNode}) {
-    return shouldGetNode(this) ? this : null;
-  }
-
-  @override
-  FileNode? visitNode({required Predicate shouldGetNode}) {
-    return shouldGetNode(this) ? this : null;
   }
 
   @override